Rework Timer module more HW Accurately
Implement a System Tick that is always counting, which informs both DIV
and TIMA changing. This means moving control of the DIV register into
the timer module. This is accurate to how the hardware works and fixes
issues caused by inaccurate timer control.

Note: There are still known glitches to implement involving interrupt
timing and glitching of TIMA when writing to DIV and such.